Move services to status-node
Move all the services to status-node, as some of them were there, some
of them were in the geth backend and scattered around.

+++ api/geth_backend.go
@@ -11,18 +11,13 @@ import (
 	"sync"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/ethereum/go-ethereum/event"
 	"github.com/ethereum/go-ethereum/log"
-	gethnode "github.com/ethereum/go-ethereum/node"
 	signercore "github.com/ethereum/go-ethereum/signer/core"
 
 	"github.com/status-im/status-go/account"
 	"github.com/status-im/status-go/appdatabase"
-	"github.com/status-im/status-go/appmetrics"
 	"github.com/status-im/status-go/connection"
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
@@ -32,19 +27,8 @@ import (
 	"github.com/status-im/status-go/node"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/rpc"
-	accountssvc "github.com/status-im/status-go/services/accounts"
-	appmetricsservice "github.com/status-im/status-go/services/appmetrics"
-	"github.com/status-im/status-go/services/browsers"
-	localnotifications "github.com/status-im/status-go/services/local-notifications"
-	"github.com/status-im/status-go/services/mailservers"
-	"github.com/status-im/status-go/services/permissions"
 	"github.com/status-im/status-go/services/personal"
-	"github.com/status-im/status-go/services/rpcfilters"
-	"github.com/status-im/status-go/services/rpcstats"
-	"github.com/status-im/status-go/services/subscriptions"
 	"github.com/status-im/status-go/services/typeddata"
-	"github.com/status-im/status-go/services/wakuext"
-	"github.com/status-im/status-go/services/wallet"
 	"github.com/status-im/status-go/signal"
 	"github.com/status-im/status-go/transactions"
 )
@@ -58,8 +42,6 @@ var (
 	ErrWhisperClearIdentitiesFailure = errors.New("failed to clear whisper identities")
 	// ErrWhisperIdentityInjectionFailure injecting whisper identities has failed.
 	ErrWhisperIdentityInjectionFailure = errors.New("failed to inject identity into Whisper")
-	// ErrWakuClearIdentitiesFailure clearing whisper identities has failed.
-	ErrWakuClearIdentitiesFailure = errors.New("failed to clear waku identities")
 	// ErrWakuIdentityInjectionFailure injecting whisper identities has failed.
 	ErrWakuIdentityInjectionFailure = errors.New("failed to inject identity into waku")
 	// ErrUnsupportedRPCMethod is for methods not supported by the RPC interface
@@ -81,7 +63,6 @@ type GethStatusBackend struct {
 	appDB                *sql.DB
 	statusNode           *node.StatusNode
 	personalAPI          *personal.PublicAPI
-	rpcFilters           *rpcfilters.Service
 	multiaccountsDB      *multiaccounts.Database
 	account              *multiaccounts.Account
 	accountManager       *account.GethManager
@@ -91,6 +72,7 @@ type GethStatusBackend struct {
 	selectedAccountKeyID string
 	log                  log.Logger
 	allowAllRPC          bool // used only for tests, disables api method restrictions
+
 }
 
 // NewGethStatusBackend create a new GethStatusBackend instance
@@ -101,14 +83,12 @@ func NewGethStatusBackend() *GethStatusBackend {
 	accountManager := account.NewGethManager()
 	transactor := transactions.NewTransactor()
 	personalAPI := personal.NewAPI()
-	rpcFilters := rpcfilters.New(statusNode)
 
 	return &GethStatusBackend{
 		statusNode:     statusNode,
 		accountManager: accountManager,
 		transactor:     transactor,
 		personalAPI:    personalAPI,
-		rpcFilters:     rpcFilters,
 		log:            log.New("package", "status-go/api.GethStatusBackend"),
 	}
 }
@@ -142,10 +122,12 @@ func (b *GethStatusBackend) IsNodeRunning() bool {
 func (b *GethStatusBackend) StartNode(config *params.NodeConfig) error {
 	b.mu.Lock()
 	defer b.mu.Unlock()
+	log.Info("STARTING NODE")
 	if err := b.startNode(config); err != nil {
 		signal.SendNodeCrashed(err)
 		return err
 	}
+	log.Info("STARTED NODE")
 	return nil
 }
 
@@ -167,6 +149,7 @@ func (b *GethStatusBackend) OpenAccounts() error {
 		return err
 	}
 	b.multiaccountsDB = db
+	b.statusNode.SetMultiaccountsDB(db)
 	return nil
 }
 
@@ -251,6 +234,7 @@ func (b *GethStatusBackend) ensureAppDBOpened(account multiaccounts.Account, pas
 		b.log.Error("failed to initialize db", "err", err)
 		return err
 	}
+	b.statusNode.SetAppDB(b.appDB)
 	return nil
 }
 
@@ -503,14 +487,17 @@ func (b *GethStatusBackend) StartNodeWithAccountAndConfig(
 	nodecfg *params.NodeConfig,
 	subaccs []accounts.Account,
 ) error {
+	log.Info("STARTING 1 NODE")
 	err := b.SaveAccount(account)
 	if err != nil {
 		return err
 	}
+	log.Info("STARTING 2 NODE")
 	err = b.ensureAppDBOpened(account, password)
 	if err != nil {
 		return err
 	}
+	log.Info("STARTING 3 NODE")
 	err = b.saveAccountsAndSettings(settings, nodecfg, subaccs)
 	if err != nil {
 		return err
@@ -565,66 +552,6 @@ func (b *GethStatusBackend) GetNodeConfig() (*params.NodeConfig, error) {
 	return b.loadNodeConfig()
 }
 
-func (b *GethStatusBackend) rpcFiltersService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return rpcfilters.New(b.statusNode), nil
-	}
-}
-
-func (b *GethStatusBackend) subscriptionService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return subscriptions.New(func() *rpc.Client { return b.statusNode.RPCPrivateClient() }), nil
-	}
-}
-
-func (b *GethStatusBackend) rpcStatsService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return rpcstats.New(), nil
-	}
-}
-
-func (b *GethStatusBackend) accountsService(accountsFeed *event.Feed) gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return accountssvc.NewService(accounts.NewDB(b.appDB), b.multiaccountsDB, b.accountManager.Manager, accountsFeed), nil
-	}
-}
-
-func (b *GethStatusBackend) browsersService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return browsers.NewService(browsers.NewDB(b.appDB)), nil
-	}
-}
-
-func (b *GethStatusBackend) permissionsService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return permissions.NewService(permissions.NewDB(b.appDB)), nil
-	}
-}
-
-func (b *GethStatusBackend) mailserversService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return mailservers.NewService(mailservers.NewDB(b.appDB)), nil
-	}
-}
-
-func (b *GethStatusBackend) appmetricsService() gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return appmetricsservice.NewService(appmetrics.NewDB(b.appDB)), nil
-	}
-}
-
-func (b *GethStatusBackend) walletService(network uint64, accountsFeed *event.Feed) gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return wallet.NewService(wallet.NewDB(b.appDB, network), accountsFeed), nil
-	}
-}
-
-func (b *GethStatusBackend) localNotificationsService(network uint64) gethnode.ServiceConstructor {
-	return func(*gethnode.ServiceContext) (gethnode.Service, error) {
-		return localnotifications.NewService(b.appDB, network), nil
-	}
-}
-
 func (b *GethStatusBackend) startNode(config *params.NodeConfig) (err error) {
 	defer func() {
 		if r := recover(); r != nil {
@@ -644,26 +571,12 @@ func (b *GethStatusBackend) startNode(config *params.NodeConfig) (err error) {
 	if err := config.Validate(); err != nil {
 		return err
 	}
-	accountsFeed := &event.Feed{}
-	services := []gethnode.ServiceConstructor{}
-	services = appendIf(config.UpstreamConfig.Enabled, services, b.rpcFiltersService())
-	services = append(services, b.subscriptionService())
-	services = append(services, b.rpcStatsService())
-	services = append(services, b.appmetricsService())
-	services = appendIf(b.appDB != nil && b.multiaccountsDB != nil, services, b.accountsService(accountsFeed))
-	services = appendIf(config.BrowsersConfig.Enabled, services, b.browsersService())
-	services = appendIf(config.PermissionsConfig.Enabled, services, b.permissionsService())
-	services = appendIf(config.MailserversConfig.Enabled, services, b.mailserversService())
-	services = appendIf(config.WalletConfig.Enabled, services, b.walletService(config.NetworkID, accountsFeed))
-	// We ignore for now local notifications flag as users who are upgrading have no mean to enable it
-	services = append(services, b.localNotificationsService(config.NetworkID))
 
 	manager := b.accountManager.GetManager()
 	if manager == nil {
 		return errors.New("ethereum accounts.Manager is nil")
 	}
 	if err = b.statusNode.StartWithOptions(config, node.StartOptions{
-		Services: services,
 		// The peers discovery protocols are started manually after
 		// `node.ready` signal is sent.
 		// It was discussed in https://github.com/status-im/status-go/pull/1333.
@@ -672,13 +585,7 @@ func (b *GethStatusBackend) startNode(config *params.NodeConfig) (err error) {
 	}); err != nil {
 		return
 	}
-	if config.WalletConfig.Enabled {
-		walletService, err := b.statusNode.WalletService()
-		if err != nil {
-			return err
-		}
-		walletService.SetClient(b.statusNode.RPCClient().Ethclient())
-	}
+
 	signal.SendNodeStarted()
 
 	b.transactor.SetNetworkID(config.NetworkID)
@@ -690,14 +597,6 @@ func (b *GethStatusBackend) startNode(config *params.NodeConfig) (err error) {
 		return
 	}
 	b.log.Info("Handlers registered")
-
-	if st, err := b.statusNode.StatusService(); err == nil {
-		st.SetAccountManager(b.accountManager)
-	}
-
-	if st, err := b.statusNode.PeerService(); err == nil {
-		st.SetDiscoverer(b.StatusNode())
-	}
 
 	// Handle a case when a node is stopped and resumed.
 	// If there is no account selected, an error is returned.
@@ -796,7 +695,7 @@ func (b *GethStatusBackend) SendTransaction(sendArgs transactions.SendTxArgs, pa
 		return
 	}
 
-	go b.rpcFilters.TriggerTransactionSentToUpstreamEvent(hash)
+	go b.statusNode.RPCFiltersService().TriggerTransactionSentToUpstreamEvent(hash)
 
 	return
 }
@@ -807,7 +706,7 @@ func (b *GethStatusBackend) SendTransactionWithSignature(sendArgs transactions.S
 		return
 	}
 
-	go b.rpcFilters.TriggerTransactionSentToUpstreamEvent(hash)
+	go b.statusNode.RPCFiltersService().TriggerTransactionSentToUpstreamEvent(hash)
 
 	return
 }
@@ -964,10 +863,7 @@ func (b *GethStatusBackend) ConnectionChange(typ string, expensive bool) {
 	b.log.Info("Network state change", "old", b.connectionState, "new", state)
 
 	b.connectionState = state
-	err := b.statusNode.ConnectionChanged(state)
-	if err != nil {
-		b.log.Error("failed to notify of connection changed", "err", err)
-	}
+	b.statusNode.ConnectionChanged(state)
 
 	// logic of handling state changes here
 	// restart node? force peers reconnect? etc
@@ -990,54 +886,18 @@ func (b *GethStatusBackend) AppStateChange(state string) {
 }
 
 func (b *GethStatusBackend) StopLocalNotifications() error {
-	localPN, err := b.statusNode.LocalNotificationsService()
-	if err != nil {
-		b.log.Error("Retrieving of LocalNotifications service failed on StopLocalNotifications", "error", err)
+	if b.statusNode == nil {
 		return nil
 	}
-
-	if localPN.IsStarted() {
-		err = localPN.Stop()
-		if err != nil {
-			b.log.Error("LocalNotifications service stop failed on StopLocalNotifications", "error", err)
-			return nil
-		}
-	}
-
-	return nil
+	return b.statusNode.StopLocalNotifications()
 }
 
 func (b *GethStatusBackend) StartLocalNotifications() error {
-	localPN, err := b.statusNode.LocalNotificationsService()
-
-	if err != nil {
-		b.log.Error("Retrieving of local notifications service failed on StartLocalNotifications", "error", err)
+	if b.statusNode == nil {
 		return nil
 	}
+	return b.statusNode.StartLocalNotifications()
 
-	wallet, err := b.statusNode.WalletService()
-	if err != nil {
-		b.log.Error("Retrieving of wallet service failed on StartLocalNotifications", "error", err)
-		return nil
-	}
-
-	if !localPN.IsStarted() {
-		err = localPN.Start(b.statusNode.Server())
-
-		if err != nil {
-			b.log.Error("LocalNotifications service start failed on StartLocalNotifications", "error", err)
-			return nil
-		}
-	}
-
-	err = localPN.SubscribeWallet(wallet.GetFeed())
-
-	if err != nil {
-		b.log.Error("LocalNotifications service could not subscribe to wallet on StartLocalNotifications", "error", err)
-		return nil
-	}
-
-	return nil
 }
 
 // Logout clears whisper identities.
@@ -1061,34 +921,11 @@ func (b *GethStatusBackend) Logout() error {
 
 // cleanupServices stops parts of services that doesn't managed by a node and removes injected data from services.
 func (b *GethStatusBackend) cleanupServices() error {
-	wakuService, err := b.statusNode.WakuService()
-	switch err {
-	case node.ErrServiceUnknown: // Waku was never registered
-	case nil:
-		if err := wakuService.DeleteKeyPairs(); err != nil {
-			return fmt.Errorf("%s: %v", ErrWakuClearIdentitiesFailure, err)
-		}
-		b.selectedAccountKeyID = ""
-	default:
-		return err
+	b.selectedAccountKeyID = ""
+	if b.statusNode == nil {
+		return nil
 	}
-
-	if b.statusNode.Config().WalletConfig.Enabled {
-		wallet, err := b.statusNode.WalletService()
-		switch err {
-		case node.ErrServiceUnknown:
-		case nil:
-			if wallet.IsStarted() {
-				err = wallet.Stop()
-				if err != nil {
-					return err
-				}
-			}
-		default:
-			return err
-		}
-	}
-	return nil
+	return b.statusNode.Cleanup()
 }
 
 func (b *GethStatusBackend) closeAppDB() error {
@@ -1135,9 +972,6 @@ func (b *GethStatusBackend) GetActiveAccount() (*multiaccounts.Account, error) {
 
 	return b.account, nil
 }
-func (b *GethStatusBackend) WakuExtService() (*wakuext.Service, error) {
-	return b.statusNode.WakuExtService()
-}
 
 func (b *GethStatusBackend) injectAccountsIntoServices() error {
 	chatAccount, err := b.accountManager.SelectedChatAccount()
@@ -1152,11 +986,9 @@ func (b *GethStatusBackend) injectAccountsIntoServices() error {
 		return err
 	}
 
-	wakuService, err := b.statusNode.WakuService()
+	wakuService := b.statusNode.WakuService()
 
-	switch err {
-	case node.ErrServiceUnknown: // Waku was never registered
-	case nil:
+	if wakuService != nil {
 		if err := wakuService.DeleteKeyPairs(); err != nil { // err is not possible; method return value is incorrect
 			return err
 		}
@@ -1164,29 +996,21 @@ func (b *GethStatusBackend) injectAccountsIntoServices() error {
 		if err != nil {
 			return ErrWakuIdentityInjectionFailure
 		}
-	default:
-		return err
-	}
+		st := b.statusNode.WakuExtService()
 
-	if wakuService != nil {
-		st, err := b.statusNode.WakuExtService()
-		if err != nil {
-			return err
+		if st != nil {
+			if err := st.InitProtocol(identity, b.appDB, b.multiaccountsDB, acc, logutils.ZapLogger()); err != nil {
+				return err
+			}
+			// Set initial connection state
+			st.ConnectionChanged(b.connectionState)
 		}
 
-		if err := st.InitProtocol(identity, b.appDB, b.multiaccountsDB, acc, logutils.ZapLogger()); err != nil {
-			return err
-		}
-
-		// Set initial connection state
-		st.ConnectionChanged(b.connectionState)
 	}
 
-	wakuV2Service, err := b.statusNode.WakuV2Service()
+	wakuV2Service := b.statusNode.WakuV2Service()
 
-	switch err {
-	case node.ErrServiceUnknown: // WakuV2 was never registered
-	case nil:
+	if wakuV2Service != nil {
 		if err := wakuV2Service.DeleteKeyPairs(); err != nil { // err is not possible; method return value is incorrect
 			return err
 		}
@@ -1194,15 +1018,7 @@ func (b *GethStatusBackend) injectAccountsIntoServices() error {
 		if err != nil {
 			return ErrWakuIdentityInjectionFailure
 		}
-	default:
-		return err
-	}
-
-	if wakuV2Service != nil {
-		st, err := b.statusNode.WakuV2ExtService()
-		if err != nil {
-			return err
-		}
+		st := b.statusNode.WakuV2ExtService()
 
 		if err := st.InitProtocol(identity, b.appDB, b.multiaccountsDB, acc, logutils.ZapLogger()); err != nil {
 			return err
@@ -1210,13 +1026,6 @@ func (b *GethStatusBackend) injectAccountsIntoServices() error {
 	}
 
 	return nil
-}
-
-func appendIf(condition bool, services []gethnode.ServiceConstructor, service gethnode.ServiceConstructor) []gethnode.ServiceConstructor {
-	if !condition {
-		return services
-	}
-	return append(services, service)
 }
 
 // ExtractGroupMembershipSignatures extract signatures from tuples of content/signature
